md: No split IO for ICE based dm-req-crypt
dm-req-crypt works in transparent encryption mode when ICE is used.
dm-req-crypt registers to splits the IO without checking if it is
in transparent mode or crypto mode. This causes performance degradation
for requests with size more than 128K. Adding check that would not
register for split IO if ICE is being used.

Change-Id: Id3bac7673265d940371f3c613a03e781a087956e
Signed-off-by: Dinesh K Garg <dineshg@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>